Rebased on the new version of CLI
