+++ frictionless/program/__init__.py
@@ -1,8 +1,6 @@
 from .api import program_api
-from .main import program
 from .describe import program_describe
 from .extract import program_extract
+from .main import program, program_main
 from .transform import program_transform
 from .validate import program_validate
-
-# TODO: rebase on Typer - https://github.com/tiangolo/typer
