+++ frictionless/program/api.py
@@ -1,17 +1,15 @@
-import click
-from .main import program
+from typer import Option as Opt
 from ..system import system
+from .main import program
 from .. import config
 
 
 @program.command(name="api")
-@click.option("--port", type=int, default=config.DEFAULT_SERVER_PORT, help="Server port")
-def program_api(port):
-    """Start API
-    \f
-    API      | Usage
-    -------- | --------
-    Public   | `$ frictionless api`
+def program_api(
+    port: int = Opt(config.DEFAULT_SERVER_PORT, help="Specify server port"),
+):
+    """
+    Start API server
     """
     server = system.create_server("api")
     server.listen(port=port)
