+++ frictionless/program/transform.py
@@ -1,19 +1,23 @@
-import click
+import typer
 from ..transform import transform
+from typer import Argument as Arg
 from .main import program
 
 
 @program.command(name="transform")
-@click.argument("source", type=click.Path(), required=True)
-def program_transform(source):
-    """Transform data
-    \f
-    API      | Usage
-    -------- | --------
-    Public   | `$ frictionless transform`
-    """
+def program_transform(
+    source: str = Arg(..., help="Path to a transform pipeline"),
+):
+    """Transform data source using a provided pipeline."""
+
+    # Transform source
     try:
         transform(source)
     except Exception as exception:
-        click.secho(str(exception), err=True)
-        exit(1)
+        typer.secho(str(exception), err=True, fg=typer.colors.RED, bold=True)
+        raise typer.Exit(1)
+
+    # Return default
+    typer.secho("---")
+    typer.secho(f'success: "{source}"', bold=True)
+    typer.secho("---")
