--- frictionless/program2/__init__.py
@@ -1,6 +1,0 @@
-from .api import program_api
-from .describe import program_describe
-from .extract import program_extract
-from .main import program, program_main
-from .transform import program_transform
-from .validate import program_validate
